refactor threading logic into multiplexer away from lepton processing logic
